requirements.txt
torch
torchvision
transformers
Pillow
numpy
matplotlib
tqdm
python-dotenv
roboflow
